fix: Cancel crash reports after flushing startup crashes
Internally FileStore retains a Collection of files which have previously been found for a request,
to prevent duplicate reports being sent. This puts the onus on the calling code to cancel or delete
the files once the request has been completed. This change ensures that the storedFiles are only
requested once, then filtered for crashes on launch, and then all files are cancelled after a
reasonable amount of time has passed regardless of delivery, which reset the internal FileStore
state. flushAsync is then called which will flush any remaining reports that have not yet been
delivered