test: cancel previous runnables posted to Async
This change cancels runnables in the teardown for each test. The current hypothesis is that these
are posted primarily in Client construction, which can result in flaky behaviour for tests which
touch the filesystem etc. This is intended to confirm whether this is the case, rather than as a
long-term solution