fix(ndk): Manually compute orientation value
For reasons not entirely clear, sometimes after the orientation
notification is fired, the Android content Resources may still return
the old value. This fixes the problem by computing a string value
manually from the listener degree value.

The adb commands to rotate the device don't seem to be suitable for
close-to-launch events and produce inconsistent results in the
test suite :(