tests: Add assertion for how class names resolve
When there is a class name, it should be appended into the method. This
is the default behavior as there are generally class names for methods.
In the case of native stacktraces from C/C++, the class name is empty,
so it should be excluded from the trace without a leading ‘.’