fix(signal): Improve odds of a full stack on ARM32
When a signal is raised on 32-bit ARM, the current context is the signal
stack rather than the crash stack. To work around this, set the register
state before unwinding the first frame (using the program counter as the
first frame). Then the stack can be unwound normally.

In cases where the signal code is TKILL (6), libunwind hangs / crashes
immediately after setting the register values. For these cases, the
first frame is recorded and the stack is terminated early to prevent
losing the report altogether. Alternative approaches would be to grab
every word from the stack and inspect for symbols or to bundle a
different unwinder altogether (libstackunwind maybe?).

For stack buffer overflows (and potentially other variants of
SEGV_MAPERR), libunwind crashes immediately after setting the register
values as well.