fix: Only set stack registers / exit unwind early on 32-bit
Uses the ABI list to conditionally switch unwinders on ARM.

While the stack enhancements added in #378 are intended only to be run
on 32-bit devices and gated to the __arm__ build, when 64-bit devices
are running the 32-bit library, the stacktraces for those devices were
unintentionally affected. This is the case when an app is configured to
only run 32-bit binaries (Android cannot run both 64-bit and 32-bit
binaries at once), and many popular libraries (e.g. React Native) only
distribute 32-bit libraries.

To further improve the stack quality on 32-bit, switches the underlying
unwinder from gnu unwind to non-gnu libunwind. Improves frame resolution
and does not have the crash in unwind when setting registers.

Fixes #382