fix: handle null values in MetaData.mergeMaps
MetaData tab values can be HashMaps, whose entries can be null. MetaData.mergeMaps seemed to assume
that the Map was always a ConcurrentHashMap, whose values cannot be null, which could lead to a NPE
when merging config.metaData and report.metaData for nullable values, such as activeScreen. Adding a
null check and conditionally adding values fixes this behaviour.