feat: Add C++ exception handler
Handles C++ exceptions in the std::set_terminate function rather than
waiting for the exception to propagate to the signal handler. Improves
the quality of stacktraces by working around issues in abort(), adds the
C++ exception type name as the error class and a description of the
thrown object as the message.