fix: Invoke the old signal handler if handling completed prior to raise
For the C++ handler in particular, once handling is complete, it 
uninstalls itself and calls the previous handler. The default previous
handler abort()s the app which would then trigger the signal handler.
This introduces a check whether handling is completed and if so, calls
the default/previous signal handler, terminating the application 
cleanly.