fix(ndk): Switch unwinders based on required async signal-safety
When capturing a stacktrace from a signal handler or C++ terminate
handler different unwinders deadlock on different architectures and/or
Android API levels. This change uses different unwinders based on
whether async signal safety is required or not, by supported API levels.

It introduces libunwindstack as the signal-safe unwinder for API 15+ on
ARM64 and x86/x64 devices. For notify()/C++ exceptions, use libunwind,
nongnu libunwind, or libcorkscrew depending on the API level.

* libcorkscrew hangs when unwinding from a C++ exception on ARM 32-bit.
* libunwind hangs on ARM64 from a signal handler
* nongnu libunwind hangs on ARM32 from a signal handler if raise() is
  called directly