feat: update START_SESSION ndk message to include handled count
the START_SESSION message is emitted when a session is started or resumed, and allows the NDK layer
to update its cached value for session information which is included in the error payload. It was
previously assumed that the handled count would always be 0 when starting a session - this is no
longer the case as sessions can be resumed, therefore the handled count is passed as part of the
message.