feat: implement BlockedThreadDetector for detecting liveness issues in a given looper
The BlockedThreadDetector posts a Runnable at regular intervals to a given thread, from a background
thread. The runnable updates a timestamp, and the background thread checks that this timestamp has
not exceeded a configured timeout threshold, which is taken to mean that the thread is blocked. If
this is the case, the delegate will be invoked once to allow the calling code to respond, which in
our case involves generating an error report.