refactor: use packageInfo.longVersionCode on API 28+, fixing usage of deprecated packageInfo.version
API 28 added an additional android:versionCodeMajor attribute, which can be used if an app has run
out of space in the android:versionCode int. The PackageInfo API we previously used has now been
deprecated and we should retrieve the version code from the longVersionCode method instead, which
contains an amalgam of the versionCode + versionCodeMajor.