test: optimise mazerunner fixture build speed
Mazerunner previously took ~3 mins to run the handled_exception.feature. The majority of this time
was spent performing a clean build and linting the example app, neither of which needed to happen.
The build now reuses existing outputs and does not lint the fixture, resulting in an execution time
of <1 minute for the scenario. This will predominantly benefit local development rather than CI.