feat: send bugsnag report when cached error file cannot be deserialised
Sends a report to bugsnag whenever it's not possible to deliver an error, ensuring that the user
receives at least some information. This new report will not be cached if delivery fails, to prevent
any serialisation bugs from looping and creating multiple reports. To support this change, we now
always attempt serialisation of the cached error. This removes the need for the Report class to have
a file, which has an added benefit of allowing us to specify Report#error as non-null.