fix: add null check to minimal error reports, preventing crash when attempting to extract info from
When a corrupted report is stored from a previous version of bugsnag-android on disk, information
required to construct a minimal error report is not present in the filename. This means the Error
will be returned as null - causing a NPE when attempting to notify. This adds a null check to fix
this scenario.