fix(ndk): Decode native report bytes as a string on the JVM side
NewStringUTF() is not safe to use unless you control the input 100% of
the time. It uses the Modified UTF-8 charset, which may exclude possible
inputs, crashing with "illegal byte" errors. This issue can be mitigated
by passing the raw byte array to the JVM layer, and then decoding the
string as UTF-8. Using the String(bytes[], Charset) constructor bypasses
possible exceptions, instead replacing unknown bytes with the
unidentified character marker.

This fixes a possible crash where a string value is added to a report
using an encoding other than ASCII or Unicode variants (such as
SHIFT-JIS, EUC-*, etc) which then cannot be decoded as UTF-8. For
example, half-width kana in EUC-JP is represented by 0x8E followed by
a value between 0xA1 and 0xDF.