fix(ndk): Avoid double free
ReleaseByteArrayElements() optionally releases the source array, which
in all cases in this library is a char array. However, the memory for
the array is only occasionally directly managed. So when releasing the
Java byte array, the correct `mode` should be specified to indicate what
should happen to the parent array:

0           copy back the content and free the elems buffer
JNI_COMMIT  copy back the content but do not free the elems buffer
JNI_ABORT   free the buffer without copying back the possible changes

`payload` in `deliverReport()` should be freed, as it was allocated. The
rest of the usages are not.