refactor(plugin-anr): Add enabled bool as proxy for install/uninstall
Re-enabling the previous handler (when the handler is the default
handler), causes a crash upon the next SIGQUIT as the default handler
crashes the process. (Though sigaction(2) seems to imply that 
SIG_DFL should be a no-op).

This change instead avoids uninstalling the handler and instead 
introduces an "enabled" flag which only writes to the ANR buffer when
set. Then the previously installed custom handlers will still be
invoked.