test: add mazerunner smoke test
This adds a smoke test feature, which runs scenarios that verify critical functionality such as
checking that unhandled exceptions are caught and reported. This will reduce the feedback loop on
CI/testing locally, as it allows us to check the notifier isn't fundamentally broken without having
to run the entire test suite.