fix: delete report file if handler returns without crashing
If the previous signal handler returns without crashing the app, this indicates that the report may
be a false positive. On Unity this can be triggered by loading an invalid asset then force-closing
the app, which triggers a bug in Unity that causes a SIGSEGV, without crashing the previous handler.
We do not wish to report these types of errors if the process has not been terminated, so should
delete the file in this scenario.