build(core): Prevent obfuscation of classes for native access
Compiling against latest bugsnag-android showed warnings for keeping
NativeInterface while Client was obfuscated and a part of its public 
API, and similarly for JsonStream and many of the classes in core.

Additionally added the Bugsnag class to make it available via the JNI 
when obfuscation is in effect (as classes are loaded via the JNI by
name).