Do not use resilient option after all
The way that the resilient driver has been written means that a NoSuchElement exception is
treated as a broken Appium session, which is not the case.  It happens a lot in Android 
tests (relatively normally) and we have higher level exception handlers to deal with it.
In this case the resilient driver’s handler is actually getting in the way.
[full ci]