fix(ndk): fix omissions and mis-sized memcpys in v4 migration
memcpy()'s last arg should be the size of the source, not destination.
In this case, the source breadcrumbs is smaller than the destination
array, since the max crumb count changed. This manifested as a frequent
but intermittent test failure in NativeStructMigration tests
[PLAT-7617].

Adjusted other instances for consistency, though some sizes were the
same.

Added tests in a new style which generates dedicated files for each
migration.