fix(ndk): support storing large ints for memory and time values
Added additional changes to event format v8 to expand the size of
version code, total memory, and duration values to fit 64-bit integers.
Technically any stored values can now be as large as ~52 bits when
delivered as that is how much can fit in a double within the current
JSON format.