build(ndk): define version scripts to avoid exporting unneeded symbols
reduce the surface of API, which expanded quite a bit with updated
unwindstack.

https://developer.android.com/ndk/guides/middleware-vendors

debug builds export additional symbols used in the tests

This change requires removal of code which re-throws types defined in
binaries other than libbugsnag-ndk, because if the type has been hidden,
then the throw invocation will trigger an immediate termination
(skipping the crash report), or if the type is custom (e.g. a subclass
of std::runtime_error, etc), then no useful message was detected anyway,
as the type information is likely not available.

Useful reading on the topic:

* https://gcc.gnu.org/wiki/Visibility (specifically "Problems with C++
  exceptions")
* https://developer.android.com/ndk/guides/common-problems#rttiexceptions_not_working_across_library_boundaries
* https://developer.android.com/ndk/guides/cpp-support#ic
* https://en.cppreference.com/w/cpp/language/definition (section on ODR)

Technically this lib should not export std:: symbols but this at least
preserves existing behavior until its re-evaluated.

[full ci]