build: update ndk revision
Bumping is required to compile with c++17 language features

Use latest LTS, per guidance for middleware vendors

> NDK versions are largely compatible with each other, but occasionally
there are changes that break compatibility. If you know that all of your
users are using the same version of the NDK, it's best to use the same
version that they do. Otherwise, use the newest version.

https://developer.android.com/ndk/guides/middleware-vendors

change requires update to root_detection to use 3-param open(2)

[full ci]