test(ndk): assert addresses resolve to correct locations
Removes the double free scenario, as the exact crashing condition is
dependent on undefined behavior, and thus can't assert an exact
location.

Refactors several existing tests to have their own files for their
native components with the intent to:

1. Make the native functions easier to find - one big file is getting
   unwieldy
2. Make line number assertions stable, avoiding changes if unrelated
   tests change

Avoids asserting on inconsistent behavior - for example
CXXStackoverflowScenario always crashes in the same function but line
info may not be available on arm32.