Event delivery with an event.isUnhandled = true should behave similarly to ExceptionHandler.uncaughtException's delivery
In the case where client.notify is called and the event is modified to be treated as unhandled:
```java
client.notify(throwable, new OnErrorCallback() {
      @Override public boolean onError(@NonNull Event event) {
        event.setUnhandled(true);
       ...
      }
    });
```
we should cache the event, similar to how it is done for an unhandled exception. With an unhandled exception, the app's availability will cease - and if we are marking an event as unhandled, we are indicating this will be the case.